refactor(web): rename lib/analysis-workspace/ → lib/workspace-state/
Eliminates mirrored-name confusion with components/AnalysisWorkspace/.
The new name signals "pure state models and computations, not
components." Updates all ~85 import sites plus the one relative import
in workers/analysis.worker.ts and docs in styles/README.md.
