+++ packages/dbt-tools/web/src/lib/workspace-state/executionTypeBarLabels.ts
